requirements.txt
absl-py==2.1.0
array-record==0.4.1
astunparse==1.6.3
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dm-tree==0.1.8
etils==1.7.0
flatbuffers==24.3.25
fsspec==2024.3.1
gast==0.4.0
google-auth==2.29.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.63.0
h5py==3.11.0
idna==3.7
importlib_resources==6.4.0
keras==2.10.0
Keras-Preprocessing==1.1.2
libclang==18.1.1
Markdown==3.6
MarkupSafe==2.1.5
numpy==1.26.4
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==24.0
promise==2.3
protobuf==3.20.3
psutil==5.9.8
pyasn1==0.6.0
pyasn1_modules==0.4.0
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-datasets==4.9.3
tensorflow-estimator==2.10.0
tensorflow-gpu==2.10.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow-metadata==1.15.0
termcolor==2.4.0
toml==0.10.2
tqdm==4.66.4
typing_extensions==4.11.0
urllib3==2.2.1
Werkzeug==3.0.3
wrapt==1.16.0
zipp==3.18.1
